hall/backend/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]